+++ app/assets/javascripts/main.js
@@ -12,12 +12,10 @@
 //= require jquery.charcount
 //= require jquery-placeholder
 //= require rails-timeago
-//= require browser_detection
 //= require jquery.events.input
 //= require jakobmattsson-jquery-elastic
 //= require jquery.mentionsInput
 //= require jquery.infinitescroll-custom
-//= require jquery.autocomplete-custom
 //= require jquery-ui/core
 //= require jquery-ui/widget
 //= require jquery-ui/mouse
@@ -35,6 +33,7 @@
 //= require markdown-it-sup
 //= require highlightjs
 //= require clear-form
+//= require typeahead.js
 //= require app/app
 //= require diaspora
 //= require_tree ./helpers